feat(api): implement endpoint to create token
This endpoint will be used when user signs up or logs in. Client
authenticates using Google and passes `id_token` to the API. API
verifies the `id_token`, extracts user information from it and generates
a token. If the user doesn't exist, it creates the user and sends back
user information with the token as well.

Current implementation might have an issue:
- If the user's token is expired, and user sends a request to create a
token, current implementation will return the expired token. It should
clean up the expired token, generate a new one and return it. An issue
should be created to fix this.